add self-contained redomat install
this commit allows to install *redomat* in a self-contained way
via `make install`.

the external dependencies of *redomat* were mentioned only briefly
in the README.md. to make the runtime environment of *redomat*
more reproducable we use now virtualenv and pip. to actually
deploy *redomat* on a system via `make install` we use [Pyinstaller][1]
to compile *redomat* into an executable and self-contained form.

undecisive: if and how to vendor the external dependencies (in binary
form as .whl or something similar). the same is true for *virtualenv*,
which is some kind of internal dependency.

[1]: https://github.com/pyinstaller/pyinstaller